# docs/developer/api.yaml
...
openapi: 3.1.0
info:
  title: OpenTalk Controller Frontend API
  description: Specifies the endpoints and structure of the OpenTalk Controller Frontend API
  contact:
    name: OpenTalk Team
    email: mail@opentalk.eu
  license:
    name: EUPL-1.2
    identifier: EUPL-1.2
  version: 0.35.0-alpha
servers:
  - url: /v1
paths:
  /auth/login:
    get:
      tags:
        - "api::v1::auth"
      summary: Get the configured OIDC provider
      description: |-
        Returns the relevant information for a frontend to authenticate against the
        configured OIDC provider for the OpenTalk service.
      operationId: get_login
      responses:
        "200":
          description: Get information about the OIDC provider
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetLoginResponseBody"
        "500":
          $ref: "#/components/responses/InternalServerError"
      security: []
    post:
      tags:
        - "api::v1::auth"
      summary: "**Deprecated**: This endpoint exists only for backwards compatibility and must no longer be used."
      description: |-
        The login endpoint

        Attempt to authenticate with a provided ID token. The ID token can be
        received from an OIDC provider and contains information about the requesting
        user as well as an expiration timestamp. When a valid token with an unknown user
        is provided, a new user will be created in the database.
      operationId: post_login
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthLoginPostRequestBody"
        required: true
      responses:
        "200":
          description: "Login successful, answer contains a list of permissions"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostLoginResponseBody"
              example:
                permissions: []
        "400":
          description: The provided ID token is malformed or contains invalid claims
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
              example:
                code: invalid_claims
                message: some required attributes are missing or malformed
        "401":
          description: The provided ID token is invalid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
              example:
                code: unauthorized
                message: Authentication failed
        "500":
          $ref: "#/components/responses/InternalServerError"
      deprecated: true
      security: []
  /auth/logout:
    post:
      tags:
        - "api::v1::auth"
      summary: Triggers back-channel OIDC logout
      operationId: post_logout
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/PostLogoutRequestBody"
        required: true
      responses:
        "204":
          description: Logout successfull
        "400":
          description: Logout request invalid or logout failed
      security: []
...
  /events:
    get:
      tags:
        - "api::v1::events"
      summary: Get a list of events and exceptions
      description: |-
        The events and exceptions are sorted chronologically.

        Returns a paginated list of events and their exceptions inside the given time range
      operationId: get_events
      parameters:
        - name: time_min
          in: query
          description: Optional minimum time in which the event happens
          required: false
          schema:
            oneOf:
              - type: "null"
              - $ref: "#/components/schemas/Timestamp"
        - name: time_max
          in: query
          description: Optional maximum time in which the event happens
          required: false
          schema:
            oneOf:
              - type: "null"
              - $ref: "#/components/schemas/Timestamp"
        - name: created_before
          in: query
          description: Only query events created before this timestamp
          required: false
          schema:
            oneOf:
              - type: "null"
              - $ref: "#/components/schemas/Timestamp"
        - name: created_after
          in: query
          description: Only query events created after this timestamp
          required: false
          schema:
            oneOf:
              - type: "null"
              - $ref: "#/components/schemas/Timestamp"
        - name: invitees_max
          in: query
          description: |-
            Maximum number of invitees to return inside the event resource

            Default value is 0
          required: false
          schema:
            oneOf:
              - type: "null"
              - $ref: "#/components/schemas/PageSize"
        - name: favorites
          in: query
          description: Return only favorite events
          required: false
          schema:
            type: boolean
        - name: invite_status
          in: query
          description: Filter the events by invite status
          required: false
          schema:
            type: array
            items:
              $ref: "#/components/schemas/EventInviteStatus"
        - name: per_page
          in: query
          description: How many events to return per page
          required: false
          schema:
            oneOf:
              - type: "null"
              - $ref: "#/components/schemas/PageSize"
        - name: after
          in: query
          description: |-
            Cursor token to get the next page of events

            Returned by the endpoint if the maximum number of events per page has been hit
          required: false
          schema:
            oneOf:
              - type: "null"
              - $ref: "#/components/schemas/Cursor_GetEventsCursorData"
        - name: adhoc
          in: query
          description: |-
            Only get events that are either marked as adhoc or non-adhoc

            If present, all adhoc events will be returned when `true`, all non-adhoc
            events will be returned when `false`. If not present, all events will
            be returned regardless of their `adhoc` flag value.
          required: false
          schema:
            type:
              - boolean
              - "null"
        - name: time_independent
          in: query
          description: |-
            Only get events that are either time-independent or time-dependent

            If present, all time-independent events will be returned when `true`,
            all time-dependent events will be returned when `false`. If absent,
            all events will be returned regardless of their time dependency.
          required: false
          schema:
            type:
              - boolean
              - "null"
      responses:
        "200":
          description: List of the events and exceptions
          headers:
            link:
              schema:
                $ref: "#/components/schemas/CursorLink"
              description: Links for paging through the results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EventOrException"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
      security:
        - BearerAuth: []
    post:
      tags:
        - "api::v1::events"
      summary: Create a new event
      description: Create a new event with the fields sent in the body.
      operationId: new_event
      parameters:
        - name: suppress_email_notification
          in: query
          description: Flag to disable email notification
          required: false
          schema:
            type: boolean
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostEventsBody"
        required: true
      responses:
        "201":
          description: The event has been created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventResource"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
      security:
        - BearerAuth: []
  /events/instances:
    get:
      tags:
        - "api::v1::events::instances"
      summary: Get a list of events and instances
      description: |-
        The instances are calculated based on the RRULE of the event. If no RRULE is
        set for an event, the event itself is returned

        Returns a paginated list of events or their instances inside the given time range
      operationId: get_events_and_instances
      parameters:
        - name: time_min
          in: query
          description: Optional minimum time in which the event or instance happens
          required: false
          schema:
            oneOf:
              - type: "null"
              - $ref: "#/components/schemas/Timestamp"
        - name: time_max
          in: query
          description: Optional maximum time in which the event or instance happens
          required: false
          schema:
            oneOf:
              - type: "null"
              - $ref: "#/components/schemas/Timestamp"
        - name: created_before
          in: query
          description: Only query events created before this timestamp
          required: false
          schema:
            oneOf:
              - type: "null"
              - $ref: "#/components/schemas/Timestamp"
        - name: created_after
          in: query
          description: Only query events created after this timestamp
          required: false
          schema:
            oneOf:
              - type: "null"
              - $ref: "#/components/schemas/Timestamp"
        - name: invitees_max
          in: query
          description: |-
            Maximum number of invitees to return inside the event resource

            Default value is 0
          required: false
          schema:
            oneOf:
              - type: "null"
              - $ref: "#/components/schemas/PageSize"
        - name: instances_max
          in: query
          description: |-
            Maximum number of instances to return for each recurring event

            Default value is 0
          required: false
          schema:
            $ref: "#/components/schemas/ItemCount"
        - name: favorites
          in: query
          description: Return only favorite events
          required: false
          schema:
            type: boolean
        - name: invite_status
          in: query
          description: Filter the events by invite status
          required: false
          schema:
            type: array
            items:
              $ref: "#/components/schemas/EventInviteStatus"
        - name: per_page
          in: query
          description: How many events and instances to return per page
          required: false
          schema:
            oneOf:
              - type: "null"
              - $ref: "#/components/schemas/PageSize"
        - name: after
          in: query
          description: |-
            Cursor token to get the next page of events and instances

            Returned by the endpoint if the maximum number of events and instances per page has been hit
          required: false
          schema:
            oneOf:
              - type: "null"
              - $ref: "#/components/schemas/Cursor_GetEventsAndInstancesCursorData"
        - name: adhoc
          in: query
          description: |-
            Only get events that are either marked as adhoc or non-adhoc

            If present, all adhoc events will be returned when `true`, all non-adhoc
            events will be returned when `false`. If not present, all events will
            be returned regardless of their `adhoc` flag value.
          required: false
          schema:
            type:
              - boolean
              - "null"
        - name: time_independent
          in: query
          description: |-
            Only get events that are either time-independent or time-dependent

            If present, all time-independent events will be returned when `true`,
            all time-dependent events will be returned when `false`. If absent,
            all events will be returned regardless of their time dependency.
          required: false
          schema:
            type:
              - boolean
              - "null"
      responses:
        "200":
          description: List of events and instances successfully returned
          headers:
            link:
              schema:
                $ref: "#/components/schemas/PageLink"
              description: Links for paging through the results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetEventInstancesResponseBody"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
      security:
        - BearerAuth: []
...
  "/events/{event_id}/instances":
    get:
      tags:
        - "api::v1::events::instances"
      summary: Get a list of the instances of an event
      description: |-
        The instances are calculated based on the RRULE of the event. If no RRULE is
        set for the event, the single event instance is returned.

        If no pagination query is added, the default page size is used.
      operationId: get_event_instances
      parameters:
        - name: invitees_max
          in: query
          description: Maximum number of invitees to include inside the event
          required: false
          schema:
            oneOf:
              - type: "null"
              - $ref: "#/components/schemas/PageSize"
        - name: time_min
          in: query
          description: Minimum time of the event instances
          required: false
          schema:
            oneOf:
              - type: "null"
              - $ref: "#/components/schemas/Timestamp"
        - name: time_max
          in: query
          description: Maximum time of the event instances
          required: false
          schema:
            oneOf:
              - type: "null"
              - $ref: "#/components/schemas/Timestamp"
        - name: per_page
          in: query
          description: How many events to return per page
          required: false
          schema:
            oneOf:
              - type: "null"
              - $ref: "#/components/schemas/PageSize"
        - name: after
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/GetEventInstancesCursorData"
            description: |-
              Opaque token which represents T as a base64 string (where T is encoded using postcard)

              Used for cursor based pagination
        - name: event_id
          in: path
          description: The id of the event
          required: true
          schema:
            $ref: "#/components/schemas/EventId"
      responses:
        "200":
          description: List of event instances successfully returned
          headers:
            link:
              schema:
                $ref: "#/components/schemas/PageLink"
              description: Links for paging through the results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetEventInstancesResponseBody"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
      security:
        - BearerAuth: []
...
  /invite/verify:
    post:
      tags:
        - "api::v1::invites"
      summary: Verify an invite code
      description: Verifies the invite and returns the room url for the invite code
      operationId: verify_invite_code
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostInviteVerifyRequestBody"
        required: true
      responses:
        "200":
          description: "Invite is valid, the response body tells the room id"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostInviteVerifyResponseBody"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          description: Invalid body contents received
        "500":
          $ref: "#/components/responses/InternalServerError"
      security: []
...
  /users/me/tariff:
    get:
      tags:
        - "api::v1::users"
      summary: Get the current user tariff information.
      description: Returns the tariff information for the currently logged in user.
      operationId: get_me_tariff
      responses:
        "200":
          description: Information about the tariff of the current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TariffResource"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
      security:
        - BearerAuth: []
...
components:
  schemas:
...
    AuthLoginPostRequestBody:
      type: object
      description: "Body of a *POST* request on `/auth/login`"
      required:
        - id_token
      properties:
        id_token:
          type: string
          description: The id token to use for the login
          example: bG9yZW0gaXBzdW0sIHF1aWEgZG9sb3Igc2
...
    GetLoginResponseBody:
      type: object
      description: "Body of the response to a *GET* request on `/auth/login`"
      required:
        - oidc
      properties:
        oidc:
          $ref: "#/components/schemas/OidcProvider"
          description: Description of the OIDC provider to use for the login
...
    PageSize:
      type: integer
      format: int64
      description: A page number used for paging in requests.
      example: 30
...
    PostLoginResponseBody:
      type: object
      description: "Body of the response to a *POST* request on `/auth/login`"
      required:
        - permissions
      properties:
        permissions:
          type: array
          items:
            type: string
          description: Permissions is a set of strings that each define a permission a user has.
          uniqueItems: true
...
  responses:
...
    InternalServerError:
      description: An internal server error occurred
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
    InviteCode:
      type: http
      scheme: bearer
...